fix(cli): excessive stack event polling during deployment  (#32196)
Closes https://github.com/aws/aws-cdk/issues/32186

### Reason for this change

<!--What is the bug or use case behind this change?-->

The CLI will poll for all stack events from the beginning of time, even
though it will end up using only a fraction of them to display in the
terminal. This can cause a significant slowdown of `cdk deploy`.

### Description of changes

<!--What code changes did you make? Have you made any important design
decisions?-->

Moved the pagination to the caller side, so it can decide whether or not
to poll the next page. This is how it was before `2.167.0`


https://github.com/aws/aws-cdk/blob/7bb9203eb95fe894c0d40942ff49c782a9fec251/packages/aws-cdk/lib/api/util/cloudformation/stack-event-poller.ts#L73-L74

### Description of how you validated changes

<!--Have you added any unit tests and/or integration tests?-->

Added unit test.

### Notes

- There are several functions in the
[sdk.ts](https://github.com/aws/aws-cdk/blob/main/packages/aws-cdk/lib/api/aws-auth/sdk.ts)
that perform implicit pagination to retrieve all results. From a quick
glance, none of them seem to be misusing it though (at least with
respect to how it always was). I will investigate those functions
further in a followup PR.

### Checklist
- [x] My code adheres to the [CONTRIBUTING
GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and
[DESIGN
GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*