fix(cli): proxy is not configured correctly (#32213)
We were using the `localAddress` of the request handler options, which
is the incorrect property. Instead, create an instance of `ProxyAgent`,
passing the CA certificate and a function that resolves to the proxy
address option from the command line. If no proxy address was passed, it
defaults to the environment variables.

Tested manually using `mitmproxy` with a custom certificate. 

While testing this, I noticed that the retry strategy function was
incorrect.

Closes https://github.com/aws/aws-cdk/issues/32208

### Checklist
- [x] My code adheres to the [CONTRIBUTING
GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and
[DESIGN
GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*