chore(cli): prevent test interference
Our CLI unit tests were interfering with each other because they
were writing files from and to the current directory, which is
shared between all of them.

Solve it by making a non-writeable directory before running the
tests, so that the tests that do that start throwing errors and
we can identify them. Then fix those.