chore(cli): prevent test interference (#32270)
Our CLI unit tests were interfering with each other because they were writing files from and to the current directory, which is shared between all of them.

Solve it by making a non-writeable directory before running the tests, so that the tests that do that start throwing errors and we can identify them. Then fix those.

I tried papering over the issue by trying to create tempdirs automatically, but that started to introduce all kinds of errors in tests that were already doing tempdir management themselves, and it took too long to go and figure out what was wrong there, so I'm doing this instead.

Also in this PR:

- Switch off global `silent` mode for the tests. It's very annoying if `console.log` statements don't make it out when you're debugging.
- This PR caused a couple of lines in `init.ts` to be useless (they were there for tests), and removing those causes a lack of coverage in `init.ts`'s entirety, so add some tests for that file.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*