feat(aws-ecs-patterns): Add a feature flag to override default behavior of creating public load balancers
If this flag is not set, the default behavior for `ApplicationLoadBalancedFargateService` and `NetworkLoadBalancedFargateService` is to create a public load balancer (not changed).

If this flag is set to false, the behavior is that the load balancer will be private by default.

This is a feature flag as to keep compatibility with the old behavior.

Relevant issue: https://github.com/aws/aws-cdk/issues/32274