fix(cli): mfa code is not requested when `$AWS_PROFILE` is used (#32313)
We only passed in the `mfaCode` function if we got a profile from `--profile`, not when configured using `$AWS_PROFILE`.

Reduce a miss in the duplicated code by moving the `clientConfig` to a single initialization point.

Fixes #32312.

Unfortunately this cannot be unit tested (I cannot mock the function that I need to mock), nor integ tested because it needs human input. I'm open to ideas.

In the mean time, tested manually.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*