chore(cli): stops tests from interfering with each other (#32319)
Turn on test randomization: this causes tests to run in a random order, which will expose tests that only succeed because tests before them have left the system in a specific state (or that poison tests after them because they don't clean up).

Of course, doing this exposed that we have a number of such tests, clean those up as well.

This needed to clean up mocking a lot in the garbage collector code, because it was confusing. Also remove the side-effect mock resetting that happens whenver a `MockSdk()` is instantiated, since that caused prepared mock behavior to be cleared.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*