fix(cli): warns about missing `--no-rollback` flag that is present (#32309)
The logic on `rollback` and `!rollback` was inverted in a couple of places, causing the check to be the wrong way around and making reasoning about these options harder.

Revert the logic and do a more comprehensive test suite around these options.

Also remove a stray `debug()` statement that was left in from a previous PR, and show the stack status in the error message.

Closes #32295.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*