fix(cli): short-lived credentials are not refreshed (#32354)
The CLI immediately invokes credential *providers* to produce
*credentials*, and passes the credentials around instead of the
providers. This means that if short-lived credentials expire (like
session credentials from roles), there is no way to refresh them.

CLI calls will start to fail if that happens.

To fix this, instead of resolving providers to credentials, pass
providers around instead.

Implications for auth plugins
-------------

This widens the plugin protocol: the new plugin protocol *forced* a
translation to V3 credentials, and had no way to return V3 providers.
While it is now possible to return V3 Credential Providers from the
plugin protocol, plugin writers cannot easily take advantage of that
protocol because there have been ~8 CLI releases that only support V3
credentials and will fail at runtime of V3 providers are returned.

To support this, pass a new options argument into `getProvider()`: this
will indicate whether V3 Providers are supported or not. Plugins can
return a provider if the CLI indicates that it supports V3 providers,
and avoid doing that if the CLI indicates it won't. That way, plugins
can be rewritten to take advantage of returning V3 providers without
crashing on CLI versions `2.167.0..(this releases)`.

This also affects #32111 in which the plugin contract is being moved.

Closes #32287.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*