feat(stepfunctions-tasks): support dynamic values for Glue Job Worker Type (#32453)
### Issue # (if applicable)

Closes https://github.com/aws/aws-cdk/issues/32359

### Reason for this change

Now, I can't put the workers type dynamic from payload input in the state machine because it only accept a enum type workerType that its a enum and only accepts the define workers. If i pass a string that references the payload input it shows an error.

### Description of changes

Turn the ENUM type `WorkerType` to a class-based implementation. This should be backward compatible as there's no change on user side.

### Description of how you validated changes

New integ tests that use dynamic value for Worker type. When invoking StateMachine with the JSON payload, it will fetch the data correctly and invoke the Glue job with the current worker type and number of workers.

### Checklist
- [ ] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*