feat(route53): added EvaluateTargetHealth to Route53 Alias targets (#9481) (#30664)
### Issue # (if applicable)

Part of a bigger piece of work for #9481 
Another PR for this ticket #30739 

### Reason for this change

At this moment constructing the Route53 health checks is available only via L1 construct, this PR is one of the series to introduce the R53 health checks via higher-level constructs.

### Description of changes

The underlying L1 construct of a few L2 constructs for Alias targets already has the `EvaluateTargetHealth` property, so by adding this prop to the L2 construct we get the quick win to introduce the health checks for Alias records.

### Description of how you validated changes

- unit tests
- integration tests

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*