fix(route53-targets): deprecated method for dns name is used in userpool domain target (under feature flag) (#31403)
### Reason for this change



The [PR](https://github.com/aws/aws-cdk/pull/31402) created a new method to get CloudFront DNS name in Cognito user pool domain. A custom resource is created in the old method, but is not in the new method.

However, the `UserPoolDomainTarget` in the `route53-targets` module continues to use the old method. So we should change to use the new method.

### Description of changes



The `bind` method in the `UserPoolDomainTarget` implements `IAliasRecordTarget` interface, so a new method instead of the `bind` cannot be created.

Therefore, I take it using a feature flag.

### Description of how you validated changes



Both unit and integ tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*