chore: remove dependency on in-source compilation of `bin/` directory (#32476)
The `bin/cdk` binary relies on the fact that `bin/cdk.ts` undergoes an in-source compilation, producing `bin/cdk.js` which is then required.

This fails in a different setup where we only compile the `lib/` directory and nothing else.

So skip one level of indirection: just put the code that loads the actual CLI entry point directly in the bash wrapper.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*