chore(cdk-cli-wrapper): fix tests for monorepo
In a slightly different testing setup, the `jest.spyOn` calls fail
with the error:

```
Cannot redefine property: spawnSync
```

This has to do with how the object representing the `fs` module
is defined. We can override the module and return a plain
object with all the functions, so that the elements can be mocked
later on.