fix(cli): getting credentials via SSO fails when the region is set in the profile (#32520)
We were reading the region from the config file and passing it to the credential providers. 
However, in the case of SSO, this makes the credential provider use that region to do the 
SSO flow, which is incorrect. The region that should be used for that is the one set in the
`sso_session` section of the config file.

The long term solution is for all the logic for handling regions in the SDK itself, without
forcing consumers to know all the intricacies of all the use cases. As a mitigation for now,
we are using the non-public `parentClientConfig` while we wait for an SDK update.

Fixes https://github.com/aws/aws-cdk/issues/32510.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*