chore(yargs-gen): use lodash.clonedeep instead of structured clone
This will make it easier to run the build on node 16.

Required adjacent changes:

pkglint had a bug that made it fail for packages with a `.` in their name. Instead now allow to pass the json path directly as an array.

Make the yargs-gen script in the CLI not prefer ts files, so that it doesn't have to re-compile most of the cli code when running.