chore(yargs-gen): use lodash.clonedeep instead of structured clone (#32537)
This will make it easier to run the build on node 16.

Adjacent changes:

pkglint had a bug that made it fail for packages with a `.` in their name. Instead now allow to pass the json path directly as an array.

Move inline function extraction completely into config and not part of yargs-gen.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*