chore(cli): remove CDK_INTEG_MODE hack (#32539)
This hack was used to support the previous integration test system for the monorepo.
It was never publicly advertised and is not used in our setup anymore.
We are removing the unused code to simplify toolkit code.

### Reason for this change

Cleaning up unused code.

### Description of changes

Removed the unused code block.

### Description of how you validated changes

Searched our repo and public GitHub for any usage. Only comes up in forks of the AWS CDK.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*