chore(eks): add warning to make kubectlLayer property required in the future (#32483)
### Reason for this change

`kubectlLayer` property in EKS Cluster is currently optional. If not provided, it will use a very outdated version (1.20). This default outdated version adds a outdated dependency to CDK which should be removed.

However, we can't simply update the version because it will be a breaking change. To avoid this issue happening again, we can send a warning first then we can do a one-time breaking change to make the property required.

Another reason is the synth error (users have to provide the property) is better than we upgrading the version silently which may lead to some unexpected behavior.

### Description of changes
Update warning to let customers know this property will become required in 2025 Jan.

### Description of how you validated changes



### Checklist
- [ ] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*