chore: upgrade eslint rule plugins
In the new monorepo, we are required to use `eslint@9` because projen
forces it.

`eslint@9` requires `typescript-eslint` plugins @8, and rules have
disappeared from those plugins and been moved to the `@stylistic`
plugin.

Make the same changes in this repo already so that we can move the
code over without changing it.