chore(cli): refactor notices.ts to not use configuration settings (#32558)
This is part of a larger effort to make sure that all command-level code does not use `configuraiton.settings` directly. If it needs a CLI setting, it should be passed in as a parameter to the command-level code. The idea is that the grab bag of `configuration.settings` stays in `cli.ts` and only the relevant information is passed along to each command.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*