chore: upgrade eslint rule plugins (#32573)
In the new monorepo, we are required to use `eslint@9` because projen forces it.

* `eslint@9` in the new repo requires `typescript-eslint@8`, and rules have disappeared from those plugins and been moved to the `@stylistic` plugin.

We therefore move to `@typescript-eslint@8` and `@stylistic` in this repo already, so that we can move the code over integrally.

* `typescript-eslint@8` requires `eslint@8`, so also upgrade `eslint` to 8 in this repo.
* `eslint@8` requires an update of `eslint-plugin-jest`

This change activated a number of rules that were being silently ignored in the past, or upgrades of rule packages work in more situations. In any case, it also includes a number of code fixes to make the new rules pass.

Notable changes:

- `toThrowError` -> `toThrow`
- `toBeCalledWith` -> `toHaveBeenCalledWith`
- Insertion of a bunch of missing semicolons

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*