refactor(cli): docs, doctor, context commands do not use configuration directly anymore (#32578)
This is a refactor that follows #32558. The effort is to ensure that the `configuration` object is not directly used by commands and instead, the relevant information is passed directly to the function.

- doctor: no options needed
- docs: 1 option needed
- context: in addition to a few options, the context command also saves *some* values to `cdk.context.json`. this requires some modifications to the `Context` object to support this. 

We move away from using the archaic API in `command-api.ts`. This seems to be an old effort to standardize CLI commands that was only used for doctor/docs/context. We have since evolved to use command-specific property bags.

Since this is a refactor PR. I aim to change no functionality of the CLI.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*