chore(cli): remove unreachable v1 compat in stack selection (#32620)
### Reason for this change

Removing dead code. The version number of the `CLI is never < 2` in release and thus this code cannot be reached.
This allowed to remove the env variable hack that enabled test scenarios to use v2 behavior.

### Description of changes

Remove the unreachable if statement and then optimize code.
Also includes some minor reformatting of method parameters.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

Test pipeline.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*