chore: add feature flag report to package (#32607)
Include a file `recommended-feature-flags.json` into the `aws-cdk-lib` package. 

At build time, the CLI will copy the current `aws-cdk-lib` and `constructs` version, as well as the feature flag report, into its own package directory, and it will use those to initialize a new project.

This decouples the init version number from the CLI version number, and it removes the need for the CLI to read `cx-api` at init time.

Closes https://github.com/aws/aws-cdk-cli/issues/7

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*