chore(spec2cdk): add tsdoc comment to describe the existance of a hardcoded list of resources in spec2cdk (#32616)
### Reason for this change

Discovered this hardcoded list of resource types and did some investigation into the reasoning behind this very large hardcoded list of resources in tagging.ts. Discovered this list is non-updating list is proper and represents a perpetual state of truth for these resource. Legacy resource types often refers to resources added prior to the introduction of the Cloudformation Registry, so I added this comment to explain the purpose of this list so that future engineers will not need to perform the same investigation.

### Description of changes

added a tsdoc comment

### Description of how you validated changes

Just a comment

### Checklist
- [ ] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*