chore(cdk-build-tools): allow packing of private packages for local testing (#32644)
### Reason for this change

Sometimes there is a need to pack a private package so it can be tested as a package locally.
For example when a package is not released during development.
Previously it was not possible to pack a private package with our tooling.

### Description of changes

Adds a `--private` flag to the `cdk-package` command to allow force packing of private packages.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

Manually used the new flag to pack a private package.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*