fix(cli): CLI still fails for some plugins returning `expiration: null`
There is one more location where we need to convert `null` into
`undefined` if a plugin provider returns `expiration: null`. The SDKv3
code base also expects this value to be `undefined | Date`, just like
our own code does; when the plugin returns `null` we pass the value over
to the SDKv3 which then does a `getTime()` on the value, and then it
crashes.

Issue this fix as a patch, so that we can unblock users of a commonly
used plugin at Amazon.