fix(cli): cli still fails for some plugins returning `expiration: null` (#32668)
There is one more location where we need to convert `null` into `undefined` if a plugin provider returns `expiration: null`. The SDKv3 code base also expects this value to be `undefined | Date`, just like our own code does; when the plugin returns `null` we pass the value over to the SDKv3 which then does a `getTime()` on the value, and then it crashes.

Issue this fix as a patch, so that we can unblock users of a commonly used plugin at Amazon.

This is a PR targeting `v2-release`, picking exactly one line from #32111. After merging this we can patch-release.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*