chore: unconnected CliIoHost logger-only implementation (#32503)
### Issue #32345

Closes #32345

### Reason for this change

Setting the ground work for our [Programmatic Toolkit](https://github.com/aws/aws-cdk-rfcs/pull/654)

### Description of changes

Created an unconnected CLIIoHost with a singular initial action available `notify`. In this implementation of the soon to be defined IoHost we are only writing logs to stdout and stderr.

### Description of how you validated changes

Verified via unit testing as this is currently unconnected to the greater AWS CDK CLI

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*