chore(cli): update Command enum (#32688)
This is a stopgap -- `Command` should eventually be generated but I do not have capability to generate enums or an equivalent type to replace this yet. 

I am using `Command` as an exhaustive list to generate other code blocks; it was not used that way in the past so it hasn't needed to be updated with new commands as they've been added.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*