refactor(cli): make Deployments internal to Toolkit
Changes the contract to only optionally take a  `toolkitStackName` instead of a Deployments object.
This is in preparation to cleaning up the contract for the public.

Removes `toolkitStackName` from action options in favor of using the name from the Toolkit.
However in reality these options were not actually in use.

Also removes unused instances of `toolkitStackName` from `DeployStackOptions` and  `RollbackStackOptions`. Like with the toolkit options, these weren't not actually passed on.