chore(cli): function to turn yargs output to cliArguments (#32696)
**This PR has no impact on CLI functionality**

This PR generates a new function, `convertToCliArgs`, that takes in the output of the `yargs` configuration and turns it into a `CliArguments` type. This allows us to subsequently utilize the strongly-typed object in `cli.ts`, _but this PR does not actually do that_. 

The ultimate benefit is that now we can ensure exactly what goes into this type. We can then enforce that any property allowed to be specified via `cdk.json` is also an option in the cli. When all my refactoring is done, there will be no options that are only specified via `cdk.json`, and no options only specified via CLI.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*